+++ src/main/java/nicoAntonelli/managefy/services/UserRoleService.java
@@ -2,7 +2,6 @@ package nicoAntonelli.managefy.services;
 
 import jakarta.transaction.Transactional;
 import nicoAntonelli.managefy.entities.UserRole;
-import nicoAntonelli.managefy.entities.User;
 import nicoAntonelli.managefy.entities.UserRoleKey;
 import nicoAntonelli.managefy.repositories.UserRoleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,21 +30,37 @@ public class UserRoleService {
         return userRoleRepository.findAll();
     }
 
+    public Boolean ExistsUserRole(UserRoleKey userRoleID) {
+        return userRoleRepository.existsById(userRoleID);
+    }
+
     public UserRole GetOneUserRole(Long userID, long businessID) {
         UserRoleKey userRoleKey = new UserRoleKey(userID, businessID);
 
         Optional<UserRole> userRole = userRoleRepository.findById(userRoleKey);
         if (userRole.isEmpty()) {
-            throw new IllegalStateException("UserRole with ID: " + userRoleKey + " doesn't exists");
+            throw new IllegalStateException("Error at 'GetOneUserRole' - UserRole with ID: " + userRoleKey + " doesn't exists");
         }
 
         return userRole.get();
     }
 
     public UserRole CreateUserRole(UserRole userRole) {
-        // This 'Get one' methods throws error if entity doesn't exist
-        userService.GetOneUser(userRole.getUser().getId());
-        businessService.GetOneBusiness(userRole.getBusiness().getId());
+        // Validate user
+        Long userID = userRole.getUser().getId();
+        if (!userService.ExistsUser(userID)) {
+            throw new IllegalStateException("Error at 'CreateUserRole' - User with ID: " + userID + " doesn't exists");
+        }
+
+        // Validate business
+        Long businessID = userRole.getBusiness().getId();
+        if (!businessService.ExistsBusiness(businessID)) {
+            throw new IllegalStateException("Error at 'CreateUserRole' - Business with ID: " + businessID + " doesn't exists");
+        }
+
+        // Pre-loaded key
+        UserRoleKey userRoleKey = new UserRoleKey(userID, businessID);
+        userRole.setId(userRoleKey);
 
         return userRoleRepository.save(userRole);
     }
@@ -53,7 +68,7 @@ public class UserRoleService {
     public UserRole UpdateUserRole(UserRole userRole) {
         boolean exists = userRoleRepository.existsById(userRole.getId());
         if (!exists) {
-            throw new IllegalStateException("UserRole with ID: " + userRole.getId() + " doesn't exists");
+            throw new IllegalStateException("Error at 'UpdateUserRole' - UserRole with ID: " + userRole.getId() + " doesn't exists");
         }
 
         return userRoleRepository.save(userRole);
@@ -61,6 +76,10 @@ public class UserRoleService {
 
     public UserRoleKey DeleteUserRole(Long userRoleID, long businessID) {
         UserRoleKey userRoleKey = new UserRoleKey(userRoleID, businessID);
+        boolean exists = ExistsUserRole(userRoleKey);
+        if (!exists) {
+            throw new IllegalStateException("Error at 'DeleteUserRole' - User with ID: " + userRoleKey + " doesn't exists");
+        }
 
         userRoleRepository.deleteById(userRoleKey);
         return userRoleKey;
