+++ src/main/java/nicoAntonelli/managefy/services/UserService.java
@@ -24,10 +24,14 @@ public class UserService {
         return userRepository.findAll();
     }
 
+    public Boolean ExistsUser(Long userID) {
+        return userRepository.existsById(userID);
+    }
+
     public User GetOneUser(Long userID) {
         Optional<User> user = userRepository.findById(userID);
         if (user.isEmpty()) {
-            throw new IllegalStateException("User with ID: " + userID + " doesn't exists");
+            throw new IllegalStateException("Error at 'GetOneUser' - User with ID: " + userID + " doesn't exists");
         }
 
         return user.get();
@@ -36,12 +40,12 @@ public class UserService {
     public User GetOneUserByMail(String mail) {
         // Mail format validation (TO-DO: add regex)
         if (mail == null || mail.isBlank()) {
-            throw new IllegalStateException("Mail bad formatted: " + mail);
+            throw new IllegalStateException("Error at 'GetOneUserByMail' - Mail bad formatted: " + mail);
         }
 
         Optional<User> user = userRepository.findByMail(mail);
         if (user.isEmpty()) {
-            throw new IllegalStateException("User with mail: " + mail + " doesn't exists");
+            throw new IllegalStateException("Error at 'GetOneUserByMail' - User with mail: " + mail + " doesn't exists");
         }
 
         return user.get();
@@ -51,41 +55,44 @@ public class UserService {
         // Mail unique validation
         Optional<User> possibleUser = userRepository.findByMail(user.getMail());
         if (possibleUser.isPresent()) {
-            throw new IllegalStateException("Mail '" + user.getMail() + "' already taken");
+            throw new IllegalStateException("Error at 'CreateUser' - Mail '" + user.getMail() + "' already taken");
         }
 
+        user.setId(null);
         return userRepository.save(user);
     }
 
     public User ValidateUser(String mail, String password) {
         // Mail format validation (TO-DO: add regex)
         if (mail == null || mail.isBlank()) {
-            throw new IllegalStateException("Mail bad formatted: " + mail);
+            throw new IllegalStateException("Error at 'ValidateUser' - Mail bad formatted: " + mail);
         }
+
         // Password format validation (TO-DO: add regex)
         if (password == null || password.isBlank()) {
-            throw new IllegalStateException("Password bad formatted for the attempted mail: " + mail);
+            throw new IllegalStateException("Error at 'ValidateUser' - Password bad formatted for the attempted mail: " + mail);
         }
+
         // Correct mail & password validation (TO-DO: add encryption for password)
         User user = GetOneUserByMail(mail);
         if (!Objects.equals(mail, user.getMail()) || !Objects.equals(password, user.getPassword())) {
-            throw new SecurityException("Login error: mail or password mismatch, attempted mail: " + mail);
+            throw new SecurityException("Error at 'ValidateUser' - Mail or password mismatch, attempted mail: " + mail);
         }
 
         return user;
     }
 
     public User UpdateUser(User user) {
-        boolean exists = userRepository.existsById(user.getId());
+        boolean exists = ExistsUser(user.getId());
         if (!exists) {
-            throw new IllegalStateException("User with ID: " + user.getId() + " doesn't exists");
+            throw new IllegalStateException("Error at 'UpdateUser' - User with ID: " + user.getId() + " doesn't exists");
         }
         // Mail unique validation
         Optional<User> possibleUser = userRepository.findByMail(user.getMail());
         if (possibleUser.isPresent()) {
-            // Don't fail validation if it's the same user
+            // Only fail validation if it's not the same user
             if (!Objects.equals(possibleUser.get().getId(), user.getId())) {
-                throw new IllegalStateException("Mail '" + user.getMail() + "' already taken");
+                throw new IllegalStateException("Error at 'UpdateUser' - Mail '" + user.getMail() + "' already taken");
             }
         }
 
@@ -93,6 +100,11 @@ public class UserService {
     }
 
     public Long DeleteUser(Long userID) {
+        boolean exists = ExistsUser(userID);
+        if (!exists) {
+            throw new IllegalStateException("Error at 'DeleteUser' - User with ID: " + userID + " doesn't exists");
+        }
+
         userRepository.deleteById(userID);
         return userID;
     }
